renamed homepageupdate directory to homepage
