+++ src/js/components/homepage/EmailSignUp.jsx
