+++ src/js/components/homepage/features/PaneFeature.jsx
