+++ src/js/components/homepage/hero/AnimatedHeading.jsx
