+++ src/js/components/homepage/hero/CovidTooltip.jsx
