+++ src/js/components/homepage/hero/TotalAmount.jsx
