+++ src/js/containers/router/RouterRoutes.js
@@ -5,7 +5,7 @@
 
 import React from 'react';
 
-const HomepageUpdate = React.lazy(() => import('components/homepageUpdate/HomepageUpdate').then((comp) => comp));
+const Homepage = React.lazy(() => import('components/homepage/Homepage').then((comp) => comp));
 const SearchContainer = React.lazy(() => import('containers/search/SearchContainer').then((comp) => comp));
 const SearchContainerRedirect = React.lazy(() => import('containers/search/SearchContainer').then((module) => ({ default: module.SearchContainerRedirect })));
 const ExplorerLanding = React.lazy(() => import('components/explorer/landing/ExplorerLanding').then((comp) => comp));
@@ -42,7 +42,7 @@ const AnalystGuidePage = React.lazy(() => import('components/analystGuide/Analys
 export const routes = [
     {
         path: `/`,
-        component: HomepageUpdate,
+        component: Homepage,
         exact: true
     },
     {
